backend/requirements.txt
# Web服务与API
fastapi
uvicorn[standard]

# 网页转PDF
playwright

# 异步任务队列
celery
redis

# 文档处理
python-docx
beautifulsoup4

# 测试
pytest
